+++ pg_documentdb_core/include/collation/collation.h
@@ -41,4 +41,12 @@ IsCollationApplicable(const char *collationString)
 }
 
 
+static inline bool
+IsBsonTypeCollationAware(bson_type_t type)
+{
+	return type == BSON_TYPE_UTF8 || type == BSON_TYPE_DOCUMENT ||
+		   type == BSON_TYPE_ARRAY;
+}
+
+
 #endif